requirements.txt
google-genai 
langgraph 
pydantic 
python-dotenv